+++ core/src/test/java/org/musetest/core/tests/ValueSourceChangeListenerTests.java
@@ -237,7 +237,138 @@ public class ValueSourceChangeListenerTests
         Assert.assertFalse(notified.get());
         }
 
-    // TODO test for changes to sub sources
+    @Test
+    public void chainedEventFromSingleSubsource()
+        {
+        ValueSourceConfiguration source = ValueSourceConfiguration.forType(VariableValueSource.TYPE_ID);
+        ValueSourceConfiguration subsource = ValueSourceConfiguration.forValue("var1");
+        source.setSource(subsource);
+
+        final AtomicReference<SubsourceModificationEvent> notification = new AtomicReference<>(null);
+        ValueSourceChangeObserver listener = new ValueSourceChangeObserver()
+            {
+            @Override
+            public void subsourceModified(SubsourceModificationEvent event)
+                {
+                notification.set(event);
+                }
+            };
+
+        source.addChangeListener(listener);
+        subsource.setValue("var-changed");
+        Assert.assertNotNull(notification.get());
+        Assert.assertEquals(source, notification.get().getSource());
+        Assert.assertEquals(SubsourceModificationEvent.SubsourceClass.Single, notification.get().getModificationClass());
+        Assert.assertNotNull(notification.get().getModificationEvent());
+        Assert.assertEquals(subsource, notification.get().getModificationEvent().getSource());
+
+        // change the source and ensure we get the event on the new, but not the old
+        notification.set(null);
+        ValueSourceConfiguration subsource2 = ValueSourceConfiguration.forValue("var2");
+        source.setSource(subsource2);
+        subsource.setValue("value2");
+        Assert.assertNull(notification.get());
+        subsource2.setValue("value3");
+        Assert.assertNotNull(notification.get());
+        Assert.assertEquals(subsource2, notification.get().getModificationEvent().getSource());
+
+        // now remove the source and ensure further changes to not fire events on the former parent
+        notification.set(null);
+        source.setSource(null);
+        subsource.setValue("value3");
+        Assert.assertNull(notification.get());
+        }
+
+    @Test
+    public void chainedEventFromIndexedSubsource()
+        {
+        ValueSourceConfiguration source = ValueSourceConfiguration.forType(VariableValueSource.TYPE_ID);
+        ValueSourceConfiguration subsource = ValueSourceConfiguration.forValue("var1");
+        source.addSource(0, subsource);
+
+        final AtomicReference<SubsourceModificationEvent> notification = new AtomicReference<>(null);
+        ValueSourceChangeObserver listener = new ValueSourceChangeObserver()
+            {
+            @Override
+            public void subsourceModified(SubsourceModificationEvent event)
+                {
+                notification.set(event);
+                }
+            };
+
+        source.addChangeListener(listener);
+        subsource.setValue("var-changed");
+        Assert.assertNotNull(notification.get());
+        Assert.assertEquals(source, notification.get().getSource());
+        Assert.assertEquals(SubsourceModificationEvent.SubsourceClass.Indexed, notification.get().getModificationClass());
+        Assert.assertEquals(0, notification.get().getSubsourceIndex());
+        Assert.assertNotNull(notification.get().getModificationEvent());
+        Assert.assertEquals(subsource, notification.get().getModificationEvent().getSource());
+
+        // now test the event after we replace the source and then modify it
+        notification.set(null);
+        ValueSourceConfiguration subsource2 = ValueSourceConfiguration.forValue("var2");
+        source.replaceSource(0, subsource2);
+        subsource.setValue("ignorethis");
+        Assert.assertNull(notification.get());
+        subsource2.setValue("changed2");
+        Assert.assertEquals(subsource2, notification.get().getModificationEvent().getSource());
+
+        // now test the event after we remove the source and then modify it
+        notification.set(null);
+        source.removeSource(0);
+        subsource.setValue("changed3");
+        Assert.assertEquals(null, notification.get());
+        }
+
+    @Test
+    public void chainedEventFromNamedSubsource()
+        {
+        ValueSourceConfiguration source = ValueSourceConfiguration.forType(VariableValueSource.TYPE_ID);
+        ValueSourceConfiguration subsource = ValueSourceConfiguration.forValue("var1");
+        final String name = "name1";
+        source.addSource(name, subsource);
+
+        final AtomicReference<SubsourceModificationEvent> notification = new AtomicReference<>(null);
+        ValueSourceChangeObserver listener = new ValueSourceChangeObserver()
+            {
+            @Override
+            public void subsourceModified(SubsourceModificationEvent event)
+                {
+                notification.set(event);
+                }
+            };
+
+        source.addChangeListener(listener);
+        subsource.setValue("var-changed");
+        Assert.assertNotNull(notification.get());
+        Assert.assertEquals(source, notification.get().getSource());
+        Assert.assertEquals(SubsourceModificationEvent.SubsourceClass.Named, notification.get().getModificationClass());
+        Assert.assertEquals(name, notification.get().getSubsourceName());
+        Assert.assertNotNull(notification.get().getModificationEvent());
+        Assert.assertEquals(subsource, notification.get().getModificationEvent().getSource());
+
+        // now test the event after we re-add another source with the same name and then modify it
+        notification.set(null);
+        ValueSourceConfiguration subsource2 = ValueSourceConfiguration.forValue("var2");
+        source.addSource(name, subsource2);
+        subsource.setValue("ignorethis");
+        Assert.assertNull(notification.get());
+        subsource2.setValue("changed2");
+        Assert.assertEquals(subsource2, notification.get().getModificationEvent().getSource());
+
+        // use the replace() method to replace the source and check for correct notification
+        notification.set(null);
+        ValueSourceConfiguration subsource3 = ValueSourceConfiguration.forValue("var3");
+        source.replaceSource(name, subsource3);
+        subsource3.setValue("changed3");
+        Assert.assertEquals(subsource3, notification.get().getModificationEvent().getSource());
+
+        // remove the source and modify - should not get another notification
+        notification.set(null);
+        source.removeSource(name);
+        Assert.assertNull(notification.get());
+        }
     }